Add Binder support
To get the Binder example working the only required package is
tensorflow. The postBuild will install from GitHub so that it always
gets the HEAD release, so that if there is something broken for a bit on
PyPI it won't continue to be a problem until a new release is made.